Refactor client and server structure; update routing and remove unused components
- Changed import from "react-router-dom" to "react-router" in NotFound.jsx and related files.
- Removed tailwind.config.js from client and TypeScript templates.
- Updated vite.config.js to automatically open the server on start.
- Enhanced package.json scripts for better readability and organization.
- Renamed server package from "backend" to "server" for consistency.
- Added placeholder for controllers in server structure.
- Cleaned up API health check route and added example for future routes.
- Updated health check response in server to include a message.
- Removed unused components and files from client, including ApiMessage and DocumentationCard.
- Introduced a new Demo page to replace Home, showcasing API connection status.
