+++ cma-landing/src/components/ThemeToggle.jsx
@@ -16,7 +16,7 @@ export default function ThemeToggle() {
     return (
         <button
             onClick={() => setTheme(isDark ? "light" : "dark")}
-            className={`p-1 text-white dark:text-black bg-zinc-950 hover:bg-zinc-950/80 px-4 py-2 rounded-xl dark:bg-zinc-50 dark:hover:bg-zinc-50/80 flex items-center justify-center transition-all duration-300 hover:cursor-pointer border border-zinc-950 dark:border-zinc-50`}
+            className={`text-white dark:text-black bg-zinc-950 hover:bg-zinc-950/80 px-2 sm:px-4 py-2 rounded-xl dark:bg-zinc-50 dark:hover:bg-zinc-50/80 flex items-center justify-center transition-all duration-300 hover:cursor-pointer border border-zinc-950 dark:border-zinc-50`}
             aria-label="Toggle theme"
         >
             {isDark ? <Sun size={18} /> : <Moon size={18} />}
