+++ templates/js/client/src/components/Navigation.jsx
@@ -1,4 +1,4 @@
-import { Link } from "react-router-dom";
+import { Link } from "react-router";
 import ThemeToggle from "./ui/ThemeToggle";
 
 export default function Navigation() {
