+++ templates/js/client/src/pages/Demo.jsx
@@ -1,12 +1,20 @@
-import { useState } from "react";
-import { ArrowUpRight, Heart, Server, Palette } from "lucide-react";
-import ApiMessage from "../components/ApiMessage";
-import Person from "../components/ui/Person";
-import DocumentationCard from "../components/ui/DocumentationCard";
+import { useState, useEffect } from "react";
+import { Link } from "react-router";
+import {
+  ArrowUpRight,
+  Server,
+  Palette,
+  ChevronRight,
+  AlertCircle,
+  CheckCircle,
+  Loader2,
+} from "lucide-react";
+import axios from "axios";
+import { API_ENDPOINTS } from "../config/constants";
 
-export default function Home() {
-  const [openId, setOpenId] = useState(null);
+// Main Component
 
+export default function Demo() {
   return (
     <div className="space-y-12 my-4">
       {/* Header */}
@@ -59,7 +67,7 @@ export default function Home() {
                 GET
               </span>
               <span className="ml-3 font-mono text-gray-900 dark:text-zinc-100">
-                /api
+                /health
               </span>
             </div>
             <span className="text-sm text-gray-600 dark:text-zinc-400">
@@ -82,7 +90,7 @@ export default function Home() {
           </p>
         </div>
 
-        <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
+        <div className="flex flex-col gap-6">
           <DocumentationCard
             icon={Palette}
             title="Client Setup"
@@ -91,25 +99,13 @@ export default function Home() {
                 title: "Environment Variables",
                 code: "client/.env",
                 description:
-                  "Add your API URL, app name, and other client configuration",
+                  "Rename .env.example to .env and set your client configurations",
               },
               {
-                title: "Remove Demo Content",
-                code: "client/src/pages/Home.jsx",
+                title: "Remove Demo Component",
+                code: "client/src/pages/Demo.jsx",
                 description:
-                  "Replace this homepage content with your own landing page",
-              },
-              {
-                title: "Clean Up Components",
-                code: "client/src/components/ApiMessage.jsx",
-                description:
-                  "Remove demo components like ApiMessage, Person cards after setup",
-              },
-              {
-                title: "Delete .gitkeep files",
-                code: "client/src/**/.gitkeep",
-                description:
-                  "Delete placeholder .gitkeep files from different directories in client/src",
+                  "Delete the Demo component and remove it's route from App.jsx",
               },
             ]}
           />
@@ -122,7 +118,7 @@ export default function Home() {
                 title: "Environment Variables",
                 code: "server/.env",
                 description:
-                  "Set your MongoDB URI, JWT secret, and other server configuration",
+                  "Rename .env.example to .env and set your server configurations",
               },
               {
                 title: "Database Models",
@@ -149,36 +145,23 @@ export default function Home() {
 
       {/* Developer Info */}
       <div className="flex flex-col md:flex-row justify-between items-center gap-8 border-t-1 border-zinc-300 dark:border-zinc-700 pt-6 mt-14 max-w-5xl mx-auto">
-        <div className="flex items-center gap-4">
-          <p className="text-gray-600 dark:text-zinc-400">
-            Built with{" "}
-            <Heart className="inline-block w-5 h-5 fill-red-500 animate-bounce text-red-500" />{" "}
-            by
-          </p>
-          <Person
-            id="prasoon"
-            image="https://ik.imagekit.io/kenma/cma-cli/kenmapfp.jpeg?updatedAt=1758917892139"
-            alt="Prasoon"
-            links={{
-              github: "https://github.com/prasoonk1204",
-              twitter: "https://x.com/kenma_dev",
-              linkedin: "https://linkedin.com/in/prasoon1204",
-            }}
-            openId={openId}
-            setOpenId={setOpenId}
-          />
-          <Person
-            id="prantor"
-            image="https://ik.imagekit.io/kenma/cma-cli/akashipfp.jpg?updatedAt=1758917892027"
-            alt="Prantor"
-            links={{
-              github: "https://github.com/Prantor-Das",
-              twitter: "https://x.com/akashi_sde",
-              linkedin: "https://www.linkedin.com/in/prantor-das",
-            }}
-            openId={openId}
-            setOpenId={setOpenId}
-          />
+        <div className="flex items-center gap-1 text-gray-500 dark:text-zinc-400">
+          <span className="">Built by</span>
+          <Link
+            to="https://prasoonk.vercel.app"
+            target="_blank"
+            className="text-zinc-900 dark:text-zinc-100 hover:text-blue-500 hover:dark:text-blue-400 font-semibold text-lg"
+          >
+            Prasoon
+          </Link>
+          &
+          <Link
+            to="https://x.com/akashi_sde"
+            target="_blank"
+            className="text-zinc-900 dark:text-zinc-100 hover:text-blue-500 hover:dark:text-blue-400 font-semibold text-lg"
+          >
+            Prantor
+          </Link>
         </div>
         <div className="flex gap-4">
           <a
@@ -189,12 +172,135 @@ export default function Home() {
             Give Feedback
           </a>
           <a
-            href="mailmailto:create.mern.app.cli@gmail.com"
+            href="mailto:create.mern.app.cli@gmail.com"
             target="_blank"
             className="bg-zinc-200 hover:bg-zinc-300 dark:bg-zinc-800 hover:dark:bg-zinc-700 transition-all duration-200 px-4 py-2 rounded-xl font-semibold font-mono"
           >
             Mail Us
           </a>
+        </div>
+      </div>
+    </div>
+  );
+}
+
+// Subcomponents
+
+function ApiMessage() {
+  const [state, setState] = useState({
+    message: "",
+    loading: true,
+    error: null,
+  });
+
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        const response = await axios.get(API_ENDPOINTS.HEALTH);
+        setState({
+          message: response.data.message,
+          loading: false,
+          error: null,
+        });
+      } catch (error) {
+        console.error("API fetch error:", error);
+        setState({
+          message: "",
+          loading: false,
+          error:
+            error.response?.data?.error ||
+            error.message ||
+            "Failed to fetch from API",
+        });
+      }
+    };
+
+    fetchData();
+  }, []);
+
+  return (
+    <div className="rounded-xl border border-gray-200 dark:border-zinc-700 w-full ">
+      <div className="p-4">
+        <div className="flex items-center gap-3">
+          {state.loading && (
+            <>
+              <Loader2 className="h-5 w-5 animate-spin text-blue-600 dark:text-blue-400" />
+              <span className="text-gray-600 dark:text-zinc-400">
+                Connecting to API...
+              </span>
+            </>
+          )}
+
+          {state.error && (
+            <>
+              <AlertCircle className="h-5 w-5 text-red-500 dark:text-red-400" />
+              <div>
+                <p className="text-red-600 dark:text-red-400 font-medium">
+                  Connection Failed
+                </p>
+                <p className="text-sm text-gray-600 dark:text-zinc-400">
+                  {state.error}
+                </p>
+              </div>
+            </>
+          )}
+
+          {!state.loading && !state.error && (
+            <>
+              <CheckCircle className="h-5 w-5 text-green-500 dark:text-green-400" />
+              <div>
+                <p className="text-green-600 dark:text-green-400 font-medium">
+                  API Connected
+                </p>
+                <p className="text-sm text-gray-600 dark:text-zinc-400">
+                  RESPONSE : {state.message}
+                </p>
+              </div>
+            </>
+          )}
+        </div>
+      </div>
+    </div>
+  );
+}
+
+function DocumentationCard({ icon: Icon, title, items }) {
+  return (
+    <div className="rounded-xl border bg-white border-zinc-200 dark:bg-zinc-900/80 dark:border-zinc-800 shadow-sm hover:shadow-md transition-all duration-300">
+      <div className="p-6">
+        <div className="flex items-center gap-3 mb-4">
+          <div className="p-2 bg-zinc-100 dark:bg-zinc-800 rounded-lg">
+            <Icon className="w-6 h-6 text-zinc-700 dark:text-zinc-300" />
+          </div>
+          <h3 className="text-xl font-bold text-gray-900 dark:text-zinc-100">
+            {title}
+          </h3>
+        </div>
+
+        <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
+          {items.map((item, index) => (
+            <div
+              key={index}
+              className="flex items-start gap-3 p-3 bg-zinc-50 dark:bg-zinc-800/50 rounded-lg border border-zinc-200 dark:border-zinc-700 overflow-x-hidden"
+            >
+              <ChevronRight className="w-4 h-4 mt-0.5 flex-shrink-0 text-zinc-500 dark:text-zinc-400" />
+              <div className="flex-1">
+                <div className="flex flex-wrap items-center gap-2 mb-1">
+                  <span className="font-semibold text-sm text-gray-900 dark:text-zinc-100">
+                    {item.title}
+                  </span>
+                  {item.code && (
+                    <code className="px-2 py-1 text-xs bg-zinc-200 dark:bg-zinc-700 text-zinc-800 dark:text-zinc-200 rounded font-mono border border-zinc-300 dark:border-zinc-600">
+                      {item.code}
+                    </code>
+                  )}
+                </div>
+                <p className="text-sm text-gray-600 dark:text-zinc-400 leading-relaxed">
+                  {item.description}
+                </p>
+              </div>
+            </div>
+          ))}
         </div>
       </div>
     </div>
