+++ templates/js/package.json
@@ -8,7 +8,7 @@
   "scripts": {
     "client": "npm run dev --workspace client",
     "server": "npm run dev --workspace server",
-    "dev": "concurrently \"npm run server\" \"npm run client\"",
+    "dev": "concurrently -n server,client -c green,blue \"npm run server\" \"npm run client\"",
     "build": "npm run build --workspace client && npm run build --workspace server",
     "start": "npm run start --workspace server",
     "test": "npm run test --workspace client && npm run test --workspace server",
