+++ templates/js/server/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "backend",
+  "name": "server",
   "version": "1.0.0",
   "main": "src/server.js",
   "type": "module",
