+++ templates/js/server/src/routes/index.js
@@ -3,18 +3,12 @@ import users from "./users.js";
 
 const router = Router();
 
-// API health check
-router.get("/", (req, res) => {
-  res.json({
-    message: "🚀 MERN API is running successfully!",
-    version: "1.0.0",
-    timestamp: new Date().toISOString(),
-    endpoints: {
-      users: "/api/users",
-      health: "/health",
-    },
-  });
-});
+/*
+Example:-
+
+import test from "./test.js";
+router.use("/test", test);
+*/
 
 // Route modules
 router.use("/users", users);
