+++ templates/ts/client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "frontend",
+  "name": "client",
   "version": "1.0.0",
   "private": true,
   "type": "module",
@@ -20,7 +20,7 @@
     "lucide-react": "^0.544.0",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
-    "react-router-dom": "^6.23.1",
+    "react-router": "^7.9.3",
     "tailwindcss": "^4.1.13"
   },
   "devDependencies": {
