+++ templates/ts/client/src/Layout.tsx
@@ -1,4 +1,4 @@
-import { Outlet } from "react-router-dom";
+import { Outlet } from "react-router";
 import Navigation from "./components/Navigation";
 
 export default function Layout() {
