+++ templates/ts/server/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "backend",
+  "name": "server",
   "version": "1.0.0",
   "main": "dist/server.js",
   "types": "dist/server.d.ts",
