+++ templates/ts/server/src/routes/index.ts
@@ -1,25 +1,14 @@
-import { Router, Request, Response } from "express";
+import { Router } from "express";
 import users from "./users.js";
-import { ApiResponse } from "../types/index.js";
 
 const router = Router();
 
-// API health check
-router.get("/", (_req: Request, res: Response) => {
-  const response: ApiResponse = {
-    message: "🚀 MERN API is running successfully!",
-    data: {
-      version: "1.0.0",
-      timestamp: new Date().toISOString(),
-      endpoints: {
-        users: "/api/users",
-        health: "/health",
-      },
-    },
-  };
+/*
+Example:-
 
-  res.json(response);
-});
+import test from "./test.js";
+router.use("/test", test);
+*/
 
 // Route modules
 router.use("/users", users);
